Move SSR Relay to D8 and invert GPIO outputs
Adding a 2N5088 NPN transistor to drive the SSR.
This requires inversion of the logic levels on the output pin.
Having the transistor hanging off D3 was causing problems during
bootup because that pin doubles up as the Flash pin, used to put
the device into programming mode. D8 was the only other one left
available because D0 is not stable during startup, and therefore
not suitable for driving the SSR.