Add support for 10K NTC connected to A0 ADC input.
We are runnning inside of a PLA printed plastic enclosure.
PLA softens at around 60C, so we want to ensure we don't get
anywhere near that hot. Rough calibration of a 10K NTC shows
that we get around 18 counts increase per degree celcius with a reading
of 84 indicating 0 degrees.

The over-temp setting has been configured for 804 counts (~40C) for now.